refactor: address PR review feedback regarding naming, modularity, and cdk defaults
