+++ cdk.json
@@ -1,5 +1,5 @@
 {
-  "app": "go mod download && go run ./cdk",
+  "app": "go mod download && go run cdk/cdk.go cdk/network.go",
   "watch": {
     "include": [
       "**"
